Remove generic typing from CoordinateSystemed Add new Convertible<T> extends CoordinateSystemed interface
